This change had suddenly dissappeared from git

+++ api/routes/site.js
@@ -800,24 +800,24 @@ async function Site(router, sequelizeObjects) {
   /**
    * Delete license plate
    */
-  router.delete('/manage/licence/plates', async (req, res) => {
-    console.log('Removing ', req.body.data.id);
-    /*
-    sequelizeObjects.Plate.destroy({
-      where: {id: req.body.id,},
-    }).then(result => {
-      if (result === 1) {
-        res.status(200);
-        res.send('Removed licence plate.');
-      } else {
-        res.status(500);
-        res.send('Error removing licence plate. ');
-      }
-    });
-     */
-
-    res.status(200);
-    res.send('Removed licence plate.');
+  router.delete('/manage/licence/plates/:id', async (req, res) => {
+    const id = Number(req.params.id);
+    if (id > 0) {
+      sequelizeObjects.Plate.destroy({
+        where: {id: id},
+      }).then(result => {
+        if (result === 1) {
+          res.status(200);
+          res.send('Removed licence plate.');
+        } else {
+          res.status(500);
+          res.send('Error removing licence plate. ');
+        }
+      });
+    } else {
+      res.status(500);
+      res.send('No proper :id parameter provided for delete.');
+    }
   });
 
   /**
